Serve example directory via webpack devServer
Many ml5 examples need a running webserver (rather than the file system) to function correctly. This change exposes the "examples" directory as "/examples" and the built library as "/dist" when running "npm run start". Additionally, webpack is prompted to automatically open a browser with the directory listing of the example directory when being run.

The old library appears to have had a separate web server, running on a different port, to host the example files. I believe doing it this way (and keeping the directory structure the same between http and the file system) is slightly more easier to comprehend.