Remove @tensorflow/tfjs-node from dev dependencies (#133)
This is causing issues during install on machines (that don't have Xcode installed?). As @lindapaiste wrote in #109: We could consider (..) running the tests using the main/browser version. We will get warnings in the console about how we should be using the Node version, but I get a lot of those still even when I try to use the Node version.