Renamed depth images in result
Unified the 'visualizationImage' and 'depthMapImage' into a single 'image' property inside the result object, which contains a p5.Image of the depth map.

Also modified the examples to use this change and to name the result 'depthMap', following the standard set by other ml5 modules like the result of ml5.bodyPose being called 'poses' or ml5.handPose being called 'hands'. With this change, accessing the depth image can be done with 'depthMap.image'.