Add missing tensor cleanup
Removed the depth estimation tensor from the result object so we could handle disposing of it internally. Also tested ml5.tf.memory() on the current code and found a memory leak, which ended up being due to some segmentation tensors not being disposed. I replaced the disposal code being used here with the one used in the official tensorflow examples, which fixed the leak.