Getting started examples
The first two examples aimed at being starting points for using the model. One is simply a webcam depth estimation without any interface. The other is the same but using the mask to clear out the background.

Also made applying the segmentation mask the default for the model, since the it performs much better with it.