Fixed size mismatch bug
Bug was due to estimation being done on the source element intrinsic dimensions and not the display dimensions set by the user, leading to an unexpected output. Needed to resize the media given by the user before passing it to the models.

After some discussions in the discord, opted for resizing the input media through tensorflow.js own methods. I think this might be more performant than resizing the image in a canvas but didn't test them side by side to corroborate.