Control selfieSegmentation input size
Added a helper function in the imageUtilities that resizes an image from an HTML img/video element to the given dimensions, and returns it as a tensor. Used it in SelfieSegmentation to make sure the input is the right size and therefore the output received by the user is not of an unexpected size.