Adding tensorflow.js depth-estimation (#248)
* Included tf.js/depth-estimation & added pipeline, examples

* Better color map and normalization

* indentation clean up

* Make grayscale the default colormap

Also changed the mention of this in the examples.

* Resolve suggestions pertaining to examples

Removed console logs, the comments are clear enough without them. Also renamed the examples' <title> tag to match ml5.js format.

* Renamed depth images in result

Unified the 'visualizationImage' and 'depthMapImage' into a single 'image' property inside the result object, which contains a p5.Image of the depth map.

Also modified the examples to use this change and to name the result 'depthMap', following the standard set by other ml5 modules like the result of ml5.bodyPose being called 'poses' or ml5.handPose being called 'hands'. With this change, accessing the depth image can be done with 'depthMap.image'.

* Replace const with let in examples

Following p5.js style.

* Reuse initial segmentation to mask result

Code refactor to use bodySegmentation just one time and reuse its result for the final masking of the depth map.

* Add missing tensor cleanup

Removed the depth estimation tensor from the result object so we could handle disposing of it internally. Also tested ml5.tf.memory() on the current code and found a memory leak, which ended up being due to some segmentation tensors not being disposed. I replaced the disposal code being used here with the one used in the official tensorflow examples, which fixed the leak.

* Add mask dilation and make mask available in result

Added the dilation algorithm to the library. The level of dilation is controlled by the config option 'dilationFactor' which takes values between 0 and 10, corresponding to the amount of pixels to grow the background into the silouette. Larger dilation factors affect fps because they need longer loops to look for bounds.

Also made the mask available as a p5.Image in the result, under the name 'mask'. This mask is compatible with the p5 mask() function, so it is easy to use it to cut out the profile from the background.

Lastly, also optimized the helper function that turns imageData into p5.Image, by replacing set() and instead copying the imageData.data array into the pixels array.

* Getting started examples

The first two examples aimed at being starting points for using the model. One is simply a webcam depth estimation without any interface. The other is the same but using the mask to clear out the background.

Also made applying the segmentation mask the default for the model, since the it performs much better with it.

* Invert grayscale colormap

To match transformers.js: lighter pixels are closer to the camera, darker are farther from it.

* Add backdrop image to masked bg example

To help visualize what using the mask together with the depthMap does.

* Fixed size mismatch bug

Bug was due to estimation being done on the source element intrinsic dimensions and not the display dimensions set by the user, leading to an unexpected output. Needed to resize the media given by the user before passing it to the models.

After some discussions in the discord, opted for resizing the input media through tensorflow.js own methods. I think this might be more performant than resizing the image in a canvas but didn't test them side by side to corroborate.

* Simplified examples

Simplified existing examples and aligned them with the changes in config defaults.

* Remove console logs

Converted console logs to comments.

* Apply mask to result.data

Realized the mask and dilation was not being applied to the data array and therefore not to the getDepthAt() method. Fixed it for consistency.

* Add point cloud example

* Refactor point cloud example

Make the code a little simpler.

* Make HTML video element example

Since we already have a webcam video example, it felt redundant to have the depthEstimation-video example also use the webcam. So I modified it to instead showcase how to run depthEstimation with a video file.

* Add sourceFrame to result object

The depth estimation result now includes the exact frame of the input that was used to generate the returned estimation. This is useful for aligning the image with the estimation, especially if the model is running at a lower fps than the source video (which is most often the case)

* Add depthEstimation mesh example

Shows how to use the depth estimation result together with p5.js 3D geometry tools to build a live mesh of the webcam video.

* Use resizeImageAsTensor for size mismatch bug

Replace the code fixing the size mismatch bug by using the new function designed for that: resizeImageAsTensor.

* Improve example comments

* Create p5 2.0 examples

---------

Co-authored-by: nasif-co <nasifrincon96@gmail.com>